+++ src/utils/getMDXContent.ts
@@ -6,7 +6,9 @@ import glob from 'fast-glob'
 import { CT } from '@/constants/index'
 
 const showCT = process.env.ATOMIC_BRAND === CT
-const excludedTitles = showCT ? ['Button', 'Modal'] : ['ButtonCT', 'ModalCT']
+const excludedTitles = showCT
+  ? ['Alert', 'Button', 'Card', 'Pill', 'Modal']
+  : ['AlertCT', 'ButtonCT', 'CardCT', 'PillCT', 'ModalCT']
 
 export async function getMdxContent(source) {
   const contentGlob = `${source}/**/*.mdx`
@@ -34,6 +36,5 @@ export async function getMdxContent(source) {
       }
     })
   )
-
   return content.filter(({ data }) => !excludedTitles.includes(data.title))
 }